# sdks/sdkwork-agent-internal-sdk/openapi/sdkwork-agent-internal-api.openapi.yaml
openapi: 3.1.2
info:
  title: SDKWork Agent Internal API
  version: 0.1.0
  description: Application-ingress internal runtime API for SDKWork agent runtime hosts and trusted in-app automation.
...
servers:
  - url: '{applicationPublicIngressOrigin}'
    description: Runtime origin from topology env `SDKWORK_KERNEL_APPLICATION_PUBLIC_HTTP_URL`.
    variables:
      applicationPublicIngressOrigin:
        default: http://127.0.0.1:18280
        description: Local default from `standalone.development`; production profiles override via topology env files.
...
paths:
...
  /internal/v3/api/intelligence/runtime/sessions:
    get:
      x-sdkwork-owner: sdkwork-kernel
      x-sdkwork-api-authority: sdkwork-agent-internal-api
      tags: [intelligence]
      summary: List runtime sessions
      description: Cursor-only keyset list ordered by updatedAt/sessionId. Omit cursor for the first page and pass through pageInfo.nextCursor thereafter.
      operationId: runtime.sessions.list
      security:
        - ApiKey: []
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: internal-api
      x-sdkwork-auth-mode: ingress-token
      x-sdkwork-domain: intelligence
      x-sdkwork-resource: runtime
      x-sdkwork-tenant-scope: application
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/SdkworkTenantIdHeader'
        - $ref: '#/components/parameters/SdkworkUserIdHeader'
        - $ref: '#/components/parameters/SdkworkIdentityMacHeader'
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      x-sdkwork-owner: sdkwork-kernel
      x-sdkwork-api-authority: sdkwork-agent-internal-api
      tags: [intelligence]
      summary: Create a runtime session
      operationId: runtime.sessions.create
      security:
        - ApiKey: []
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: internal-api
      x-sdkwork-auth-mode: ingress-token
      x-sdkwork-domain: intelligence
      x-sdkwork-resource: runtime
      x-sdkwork-tenant-scope: application
      x-sdkwork-idempotent: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - $ref: '#/components/parameters/SdkworkTenantIdHeader'
        - $ref: '#/components/parameters/SdkworkUserIdHeader'
        - $ref: '#/components/parameters/SdkworkIdentityMacHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionItemResponse'
        default:
          $ref: '#/components/responses/Problem'
  /internal/v3/api/intelligence/runtime/sessions/{sessionId}:
    get:
      x-sdkwork-owner: sdkwork-kernel
      x-sdkwork-api-authority: sdkwork-agent-internal-api
      tags: [intelligence]
      summary: Retrieve a runtime session
      operationId: runtime.sessions.retrieve
      security:
        - ApiKey: []
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: internal-api
      x-sdkwork-auth-mode: ingress-token
      x-sdkwork-domain: intelligence
      x-sdkwork-resource: runtime
      x-sdkwork-tenant-scope: application
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
        - $ref: '#/components/parameters/SdkworkTenantIdHeader'
        - $ref: '#/components/parameters/SdkworkUserIdHeader'
        - $ref: '#/components/parameters/SdkworkIdentityMacHeader'
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionItemResponse'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      x-sdkwork-owner: sdkwork-kernel
      x-sdkwork-api-authority: sdkwork-agent-internal-api
      tags: [intelligence]
      summary: Delete a runtime session
      operationId: runtime.sessions.delete
      security:
        - ApiKey: []
      x-sdkwork-request-context: WebRequestContext
      x-sdkwork-api-surface: internal-api
      x-sdkwork-auth-mode: ingress-token
      x-sdkwork-domain: intelligence
      x-sdkwork-resource: runtime
      x-sdkwork-tenant-scope: application
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
        - $ref: '#/components/parameters/SdkworkTenantIdHeader'
        - $ref: '#/components/parameters/SdkworkUserIdHeader'
        - $ref: '#/components/parameters/SdkworkIdentityMacHeader'
      responses:
        '204':
          description: Session deleted
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Application ingress token. Runtime also accepts Authorization Bearer and x-sdkwork-access-token.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT ingress when SDKWORK_KERNEL_INGRESS_AUTH_MODE=jwt (HS256, RS256, or JWKS-backed keys).
  parameters:
    SessionIdPath:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
...
    SdkworkTenantIdHeader:
      name: x-sdkwork-tenant-id
      in: header
      required: false
      schema:
        type: string
      description: Caller tenant identity for session ownership in token ingress mode.
    SdkworkUserIdHeader:
      name: x-sdkwork-user-id
      in: header
      required: false
      schema:
        type: string
      description: Caller user identity for session ownership in token ingress mode.
    SdkworkIdentityMacHeader:
      name: x-sdkwork-identity-mac
      in: header
      required: false
      schema:
        type: string
      description: HMAC-SHA256 hex proof over "{tenantId}\n{userId}" keyed by the ingress token; required for non-loopback signed identity mode.
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated retry key. It is scoped by the authenticated tenant, principal, HTTP method, and route. Reusing it with a different request body returns 40901.
      schema:
        type: string
        minLength: 1
        maxLength: 255
        pattern: '^[A-Za-z0-9._:/-]+$'
...
    PageSize:
      name: page_size
      in: query
      required: false
      description: Page size for list pagination. Default 20, maximum 200.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque keyset cursor for high-volume lists. Mutually exclusive with `page`.
      schema:
        type: string
  responses:
    Problem:
      description: RFC 9457 problem detail response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    ProblemDetail:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          description: Request endpoint occurrence in the form `{METHOD} {routeTemplate}`.
        operationId:
          type: string
        code:
          type: integer
          format: int32
          description: Numeric error result code. MUST be non-zero. See API_SPEC.md §15.3.
          minimum: 40001
          maximum: 79999
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
        errorCategory:
          type: string
        retryable:
          type: boolean
    SdkWorkApiResponse:
      type: object
      additionalProperties: false
      required: [code, data, traceId]
      properties:
        code:
          type: integer
          format: int32
          description: Numeric success result code. MUST be 0 on HTTP 2xx JSON bodies.
          enum: [0]
          default: 0
          minimum: 0
          maximum: 0
        data:
          description: Operation-specific payload.
        traceId:
          type: string
          format: uuid
          description: Server-owned request correlation id.
    SdkWorkPageData:
      type: object
      additionalProperties: false
      required: [items, pageInfo]
      properties:
        items:
          type: array
          items: {}
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
    SdkWorkResourceData:
      type: object
      additionalProperties: false
      required: [item]
      properties:
        item:
          description: Typed domain resource for the operation.
    PageInfo:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          type: string
          enum: [offset, cursor]
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: string
          pattern: '^[0-9]+$'
        totalPages:
          type: integer
          minimum: 0
        nextCursor:
          type: [string, 'null']
        hasMore:
          type: boolean
...
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [agentId]
      properties:
        agentId:
          type: string
          minLength: 1
          maxLength: 256
        tenantId:
          type: string
        userRef:
          type: string
        model:
          type: string
          minLength: 1
          maxLength: 256
        modelProvider:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 1024
        goal:
          type: string
          minLength: 1
          maxLength: 262144
        instructions:
          type: string
          minLength: 1
          maxLength: 262144
        cwd:
          type: string
          minLength: 1
          maxLength: 4096
        workspaceRoots:
          type: array
          maxItems: 32
          items:
            type: string
            minLength: 1
            maxLength: 4096
        source:
          type: string
        kind:
          type: string
        timeoutMs:
          type: integer
          minimum: 0
          description: Session timeout in milliseconds.
        metadata:
          type: object
          maxProperties: 64
          additionalProperties:
            type: string
            maxLength: 4096
    SessionResponse:
      type: object
      required: [sessionId, source, kind, agentId, state, workspaceRoots, tokenUsage, messageCount, toolCallCount, compressionCount, changeSummary, childSessionIds, metadata]
      properties:
        sessionId:
          type: string
        source:
          type: string
        kind:
          type: string
        agentId:
          type: string
        userRef:
          type: string
          nullable: true
        tenantId:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        goal:
          type: string
          nullable: true
        state:
          type: string
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        modelProvider:
          type: string
          nullable: true
        cwd:
          type: string
          nullable: true
        workspaceRoots:
          type: array
          items:
            type: string
        instructions:
          type: string
          nullable: true
        tokenUsage:
          $ref: '#/components/schemas/TokenUsage'
        messageCount:
          type: integer
        toolCallCount:
          type: integer
        compressionCount:
          type: integer
        changeSummary:
          $ref: '#/components/schemas/ChangeSummary'
        childSessionIds:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    TokenUsage:
      type: object
      required: [inputTokens, outputTokens, cachedTokens, reasoningTokens, totalTokens]
      properties:
        inputTokens:
          type: integer
          format: uint64
        outputTokens:
          type: integer
          format: uint64
        cachedTokens:
          type: integer
          format: uint64
        reasoningTokens:
          type: integer
          format: uint64
        totalTokens:
          type: integer
          format: uint64
    ChangeSummary:
      type: object
      required: [additions, deletions, filesChanged]
      properties:
        additions:
          type: integer
        deletions:
          type: integer
        filesChanged:
          type: integer
    SessionListResponse:
      description: Paginated runtime session list following SdkWorkApiResponse envelope.
      allOf:
        - $ref: '#/components/schemas/SdkWorkApiResponse'
        - type: object
          required: [data]
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/SdkWorkPageData'
                - type: object
                  required: [items]
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/SessionResponse'
...
    SessionItemResponse:
      description: Session resource SdkWorkApiResponse envelope.
      allOf:
        - $ref: '#/components/schemas/SdkWorkApiResponse'
        - type: object
          required: [data]
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/SdkWorkResourceData'
                - type: object
                  properties:
                    item:
                      $ref: '#/components/schemas/SessionResponse'
...